4_1_chiffrement_homomorphe/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
brevitas==0.10.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
concrete-ml==1.9.0
concrete-ml-extensions==0.1.9
concrete-python==2.10.0
contourpy==1.3.3
cycler==0.12.1
dependencies==2.0.1
dill==0.4.0
fastapi==0.128.0
filelock==3.20.2
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
humanfriendly==10.0
hummingbird-ml==0.4.11
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.3
jsonpickle==4.1.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
onnx==1.17.0
onnxconverter-common==1.16.0
onnxmltools==1.14.0
onnxoptimizer==0.3.13
onnxruntime==1.18.0
packaging==25.0
pandas==2.0.3
pillow==12.1.0
protobuf==5.29.5
psutil==7.2.1
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.5.0
scipy==1.16.3
shellingham==1.5.4
six==1.17.0
skl2onnx==1.19.1
skops==0.5.0
skorch==0.11.0
starlette==0.50.0
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
torch==2.3.1
tqdm==4.67.1
triton==2.3.1
typer-slim==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
unfoldNd==0.2.3
urllib3==2.6.2
uvicorn==0.40.0
xgboost==1.6.2
z3-solver==4.13.0.0
